+++ src/test/resources/projects/nodemodules/test/default/main.js
@@ -1,0 +1,2 @@
+import { toArray } from "to-array";
+console.log(toArray(document.getElementsByClassName("foo")));